Rows should now be properly striped when searching | fixes #118
Had to do a round-a-bout method of fixing this.
There isn't any fix for this in pure CSS due to the limitations of :nth-of-type disregarding classes or any other selector we could use to limit it to only visible rows.
Instead we use a .js-striped class to load it like we did before, then on load we apply a .striped class to each :visible:odd row via JS, then remove the .js-striped class.
We can now remove/add the .striped class to visible rows when searching to fix this.